services/requirements.txt
asgiref==3.7.2
auth-token==0.0.3
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
